Consolidate StreamingQueryTestHarness into shared base class
